+++ tests/test_handler.py
@@ -15,6 +15,7 @@ from zyte_api.constants import API_URL
 from scrapy_zyte_api.handler import ScrapyZyteAPIDownloadHandler
 
 from . import DEFAULT_CLIENT_CONCURRENCY, SETTINGS, UNSET, make_handler, set_env
+from .mockserver import MockServer
 
 
 @pytest.mark.parametrize(
@@ -212,3 +213,52 @@ async def test_retry_policy(
 
         actual = request_call[0].kwargs["retrying"]
         assert actual == expected
+
+
+@ensureDeferred
+async def test_stats():
+    with MockServer() as server:
+        async with make_handler({}, server.urljoin("/")) as handler:
+            scrapy_stats = handler._crawler.stats
+            assert scrapy_stats.get_stats() == {}
+
+            meta = {"zyte_api": {"foo": "bar"}}
+            request = Request("https://example.com", meta=meta)
+            await handler.download_request(request, None)
+
+            assert set(scrapy_stats.get_stats()) == {
+                f'scrapy-zyte-api/{stat}'
+                for stat in (
+                    '429',
+                    'attempts',
+                    'error_ratio',
+                    'errors',
+                    'fatal_errors',
+                    'mean_connection_seconds',
+                    'mean_response_seconds',
+                    'processed',
+                    'status_codes/200',
+                    'success_ratio',
+                    'success',
+                    'throttle_ratio',
+                )
+            }
+            for suffix, value in (
+                ('429', 0),
+                ('attempts', 1),
+                ('error_ratio', 0.0),
+                ('errors', 0),
+                ('fatal_errors', 0),
+                ('processed', 1),
+                ('status_codes/200', 1),
+                ('success_ratio', 1.0),
+                ('success', 1),
+                ('throttle_ratio', 0.0),
+            ):
+                stat = f"scrapy-zyte-api/{suffix}"
+                assert scrapy_stats.get_value(stat) == value
+            for name in ('connection', 'response'):
+                stat = f"scrapy-zyte-api/mean_{name}_seconds"
+                value = scrapy_stats.get_value(stat)
+                assert isinstance(value, float)
+                assert value > 0.0