+++ CHANGES.rst
@@ -1,6 +1,18 @@
 Changes
 =======
 
+To be released
+--------------
+
+* Requires zyte-api_ ≥ 0.3.0.
+
+* Stats from zyte-api are now copied into Scrapy stats. The
+  ``scrapy-zyte-api/request_count`` stat has been renamed to
+  ``scrapy-zyte-api/processed`` accordingly.
+
+.. _zyte-api: https://github.com/zytedata/python-zyte-api
+
+
 0.3.0 (2022-07-22)
 ------------------